+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - properly handling Always approval with tool calls
 - added /subtask command to create and switch to a subtask
 - added file watch mode setting to configure file change detection method
+- properly focusing search input in task sidebar
 
 ## [0.66.0]